chore(backend): make "Java 25, not older" a build gate instead of an audit (#329)
Adds -Xlint:deprecation,removal and -Werror to both compile executions.
Without the lint flag javac reports a deprecated call as a two-line note
carrying no file and no line, and the build passes; the property "this
codebase uses no superseded API" therefore held only for as long as somebody
kept checking by hand.

Verified by watching the gate fail, not by reading the docs: a temporary
`new Locale("cs","CZ")` in main produces

  [10,16] Locale(String,String) in java.util.Locale has been deprecated
  warnings found and -Werror specified

and exit 1. The probe was removed again.

The backend used no deprecated or removal-marked API before this either --
javac -Xlint:deprecation,removal over all 367 sources reported nothing, with
a positive control confirming the lint fires. Full -Xlint:all now leaves only
serial (18 exceptions without serialVersionUID) and restricted (7 FFM
downcalls in the DPAPI store, which the launcher already permits).

Turning on -Werror required clearing 14 Error Prone warnings that nothing
surfaced before, because the build only fails on ERROR:

  - 5x MissingSummary: javadoc opening straight into @param/@return/@throws
    got the summary sentence it was missing.
  - NotJavadoc in ImapFolderService, a real defect: the javadoc for
    findFolderNamesByRole sat orphaned above *another* method's javadoc, so
    it documented nothing and its own method had no description at all.
  - AlmostJavadoc and RedundantControlFlow in HtmlSanitizer.
  - Four deliberate findings now carrying @SuppressWarnings with the reason
    next to them: CorrespondentEntity.account is the owning side of a JPA
    mapping Hibernate reads reflectively; two MailFacadeTest fields exist for
    @InjectMocks; and `member == arrival` in ThreadingService compares
    identity on purpose, since an entity with no id has nothing to compare
    through equals.

Remaining older-API spots went with it: Collections.emptyList() to List.of()
in six places (the repo already had 31 List.of against those six), the two
new Thread(...) + setDaemon(true) pairs to Thread.ofPlatform(), the
isPresent() && get().isFresh() chain in OAuth2TokenService to
filter(CachedToken::isFresh), and the concatenated MIME literal in
MalformedBodyStructureSyncIT to a text block. That last one carries CRLF the
text block does not supply on its own, so the result is verified byte
identical to the concatenation (204 chars, equals true) -- the first attempt
at it silently dropped the \r.


Docs-impact: none — an Optional idiom rewrite in the token cache lookup; nothing about what leaves the device, where it goes or what is stored.

+++ backend/src/test/java/org/voxrox/mailbackend/feature/contact/service/ContactServiceTest.java
@@ -676,7 +676,7 @@ class ContactServiceTest {
         c.setSurname(surname);
         for (int i = 0; i < emails.length; i++) {
             ContactEmailEntity em = new ContactEmailEntity();
-            em.setId((long) (10 * id + i));
+            em.setId(10 * id + i);
             em.setEmail(emails[i]);
             em.setPrimary(i == 0);
             em.setContact(c);
